requirements.txt
certifi==2022.6.15+computecanada
charset-normalizer==2.1.0+computecanada
idna==3.3+computecanada
joblib==1.1.0+computecanada
numpy==1.23.0+computecanada
Pillow==9.2.0+computecanada
Pillow-SIMD==9.0.0.post1+computecanada
requests==2.28.1+computecanada
scikit-learn==1.1.1+computecanada
scipy==1.8.0+computecanada
sklearn==0.0+computecanada
tabulate==0.8.10+computecanada
threadpoolctl==3.1.0+computecanada
tk==0.1.0
torch==1.12.0+computecanada
torchvision==0.13.0+computecanada
tqdm==4.64.0+computecanada
typing-extensions==4.3.0+computecanada
urllib3==1.26.11+computecanada
